requirements.txt
certifi==2021.5.30
cffi==1.15.0
greenlet==1.1.2
importlib-metadata==4.8.1
pycparser==2.20
pyDistAlgo==1.1.0b15
PyNaCl==1.4.0
six==1.16.0
typing-extensions==3.10.0.2
zipp==3.6.0
